Add files via upload
EssentialsRewards is a lightweight and flexible plugin designed to enhance player engagement by automatically rewarding them for achievements and milestones. It integrates seamlessly with Essentials and supports PlaceholderAPI for dynamic placeholders. With robust storage options, including MongoDB support, and a variety of commands to manage stats, messages, and storage settings, EssentialsRewards is perfect for servers looking to add a rewarding experience for their players. Customize permissions for players and admins, and let EssentialsRewards handle the rest!